Delete 404 Error Page
This error page was the default error page for firebase. Decided to change it however I am not uploading it to GitHub mainly for security reasons, and some other code, I am pretty far ahead however when you do decide to clone my repository, you will only see a (responsive) navBar and yeah, again for security reasons. Sorry guys 😞 it is what it is.